demo_application/common_requirements.txt
requests==2.22.0
attrs==19.3.0